+++ common/src/main/java/me/drex/villagerconfig/common/util/duck/IVillager.java
@@ -1,0 +1,7 @@
+package me.drex.villagerconfig.common.util.duck;
+
+import net.minecraft.server.level.ServerLevel;
+
+public interface IVillager {
+    void villagerConfig$recordBreed(ServerLevel level);
+}