+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `maxRestocksPerDay` config option
+- `maxBreedsPerDay` config option
+- `cureTime` config option
 
 ## [4.5.4] - 2026-06-15
 ### Changed